+++ src/Dependencies/PooledObjects/IPooled.cs
