+++ src/Dependencies/PooledObjects/Microsoft.CodeAnalysis.PooledObjects.projitems
@@ -12,9 +12,11 @@
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)ArrayBuilder.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ArrayBuilder.Enumerator.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)IPooled.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ObjectPool`1.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PooledDelegates.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PooledDictionary.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)PooledDisposer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PooledHashSet.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PooledStringBuilder.cs" />
   </ItemGroup>
