+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Core/CompilerExtensions.projitems
@@ -609,9 +609,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\SymbolUsageInfo.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\TypeOrNamespaceUsageInfo.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\ValueUsageInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ObjectPools\ArrayBuilder.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ObjectPools\Extensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ObjectPools\IPooled.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ObjectPools\PooledDisposer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ObjectPools\PooledHashSet.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ObjectPools\PooledObject.cs" />
